Split updater to a separate dedicated module
Simplify `main.py` and move logic into an object which represents
the whole update process.

This makes it simpler to create and share a `VersionMap` across
functions, and reduces the number of times data has to be repeatedly
read from disk. (Still not optimal.)

It also makes it much simpler to make a test interface which is
homogenous with the CLI entrypoint.